+++ src/main/java/org/opensearch/index/store/experimentals/async_io/Helper.java
@@ -2,7 +2,7 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.opensearch.index.store.async_io;
+package org.opensearch.index.store.experimentals.async_io;
 
 import java.util.concurrent.CompletableFuture;
